Move coredata classes and helpers to their own files
