+++ MyEssentialFeed/MyEssentialFeed/Feed Cache/Infrastructure/CoreData/ManagedCache.swift
@@ -1,0 +1,33 @@
+//
+//  ManagedCache.swift
+//  MyEssentialFeed
+//
+//  Created by Tiago Chaves on 04/01/24.
+//
+
+import CoreData
+
+@objc(ManagedCache)
+class ManagedCache: NSManagedObject {
+    @NSManaged var timestamp: Date
+    @NSManaged var feed: NSOrderedSet
+}
+
+extension ManagedCache {
+    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
+        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
+        request.returnsObjectsAsFaults = false
+        return try context.fetch(request).first
+    }
+    
+    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
+        try find(in: context).map(context.delete)
+        return ManagedCache(context: context)
+    }
+    
+    var localFeed: [LocalFeedImage] {
+        return feed
+            .compactMap { $0 as? ManagedFeedImage }
+            .map { LocalFeedImage(id: $0.id, description: $0.imageDescription, location: $0.location, url: $0.url) }
+    }
+}
